+++ src/components/pages/blog/blog-lede.module.scss
@@ -1,4 +1,4 @@
-@import '../../scss/colors.scss';
+@import '../../../scss/colors.scss';
 
 .lede {
   margin: 2rem 0 2.5rem;
