+++ src/pages/index.js
@@ -9,7 +9,7 @@ import SEO from '../components/layout/seo'
 import Footer from '../components/layout/footer'
 import PressLogos from '../components/pages/homepage/press-logos'
 import PressList from '../components/common/press-list'
-import BlogList from '../components/common/blog-list'
+import BlogList from '../components/pages/blog/blog-list'
 import Visualizations from '../components/pages/homepage/visualizations'
 import homepageStyles from './index.module.scss'
